+++ CHANGELOG.md
@@ -9,8 +9,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+### Changed
+
+### Fixed
+
+### Removed
+
+## [0.1.3]
+
+### Added
+
 - Added `GroupMemberStore` to improve lookup performance
+- Added `GroupMemberStore` initialization when creating a new group
+- Added `GroupEventStore` initialization when creating a new group
+- Added `UserNotifications` initialization when creating a new profile
+-
 - Added `EventAttendeeStore` to improve loopup performance
+- Added `EventAttendeeStore` initialization when creating a new event
+-
 - Added `MemberCollection` struct for usage in `GroupMemberStore` and `EventAttendeeStore`
 
 - Added `GroupEventStore` to improve lookup performance
@@ -20,14 +36,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- non-silent notification now return a `NotificationResponse` instead of a `Notification`
-- changed silent notification logic where the `silent: bool` is replaced with an `Option<notification_id>`
+- Notification now return a `NotificationResponse` which hold additional metadata instead of a `Notification`
+- Change `id: u64` to `id: Option<u64>` in `NotificationResponse` where a non-set `id` means a silent notification
+- change `FriendRequest` enum value from `FriendRequest` to `FriendRequestResponse`
+- change `JoinEventUserRequest` enum value from `u64` to `InviteAttendeeResponse`
+- change `JoinEventOwnerRequest` enum value from `u64` to `InviteAttendeeResponse`
+- change `JoinGroupUserRequest` enum value from `u64` to `InviteMemberResponse`
+- change `JoinGroupOwnerRequest` enum value from `u64` to `InviteMemberResponse`
+- change `FriendRequestAccept` enum value from `u64` to `FriendRequestResponse`
+- change `FriendRequestDecline` enum value from `u64` to `FriendRequestResponse`
+- change `remove_notifications` to `remove_user_notifications` from `NotificationCalls`
 
 ### Fixed
 
 - Fixed `event_count` and `member_count` by usage of the new stores
 
 ### Removed
+
+- Removed `SilentNotification` and `SendSilentNotification` from `NotificationType` enum
+- Removed `get_user_notifications` query call
 
 ## [0.1.2]
 
@@ -94,6 +121,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Missing `notification_id` on migration models
 
+[0.1.3]: https://github.com/Catalyze-Software/proxy/compare/master...0.1.3
 [0.1.2]: https://github.com/Catalyze-Software/proxy/compare/master...0.1.2
 [0.1.1]: https://github.com/Catalyze-Software/proxy/compare/master...0.1.1
 [0.1.0]: https://github.com/Catalyze-Software/proxy/releases/tag/0.1.0